# api/openapi.yaml
openapi: 3.0.3
info:
  title: WhereToLive HTTP API
  version: 0.1.0
  description: Authoritative HTTP contract for the WhereToLive modular monolith.
servers:
  - url: /api
paths:
...
  /auth/refresh:
    post:
      operationId: refresh
      summary: Rotate the refresh secret and issue an access token
      tags: [authentication]
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Session rotated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          $ref: "#/components/responses/AuthenticationFailed"
        "403":
          $ref: "#/components/responses/CsrfFailed"
...
  /audit/events:
    get:
      operationId: listAuditEvents
      summary: List redacted audit events in the authenticated tenant
      tags: [audit]
      security: [{ bearerAuth: [] }]
      x-wheretolive-permission: { resource: audit.events, action: read }
      parameters:
        - name: action
          in: query
          schema: { type: string, maxLength: 128 }
        - name: resource
          in: query
          schema: { type: string, maxLength: 128 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Newest-first bounded tenant audit page
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AuditEvent" } }
        "400": { description: Invalid audit filter or page }
        "401": { $ref: "#/components/responses/AuthenticationFailed" }
        "403": { $ref: "#/components/responses/AuthorizationFailed" }
  /audit/events/{eventId}:
    parameters:
      - $ref: "#/components/parameters/EventId"
    get:
      operationId: getAuditEvent
      summary: Read one redacted audit event in the authenticated tenant
      tags: [audit]
      security: [{ bearerAuth: [] }]
      x-wheretolive-permission: { resource: audit.events, action: read }
      responses:
        "200":
          description: Redacted audit event with state snapshots
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditEvent" }
        "401": { $ref: "#/components/responses/AuthenticationFailed" }
        "403": { $ref: "#/components/responses/AuthorizationFailed" }
        "404": { description: Event not found in the tenant }
  /platform/audit-events:
    get:
      operationId: listPlatformAuditEvents
      summary: List redacted audit events across tenants as a global platform administrator
      tags: [audit]
      security:
        - platformBearerAuth: []
      parameters:
        - name: tenantId
          in: query
          schema: { type: string, format: uuid }
        - name: action
          in: query
          schema: { type: string, maxLength: 128 }
        - name: resource
          in: query
          schema: { type: string, maxLength: 128 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Newest-first bounded platform audit page
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PlatformAuditEvent" }
        "400": { description: Invalid audit filter or page }
        "401": { $ref: "#/components/responses/AuthenticationFailed" }
components:
  securitySchemes:
    consumerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    platformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
        maxLength: 256
...
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    AuthenticationFailed:
      description: Tenant or credentials are invalid, or the session is unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    CsrfFailed:
      description: CSRF validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    AuthorizationFailed:
      description: The authenticated actor lacks the required permission
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    AuditEvent:
      type: object
      additionalProperties: false
      required: [id, actorType, actorId, tenantId, action, resource, resourceId, reason, result, correlationId, occurredAt]
      properties:
        id: { type: string, format: uuid }
        actorType:
          type: string
          enum: [platform_administrator, tenant_user]
        actorId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid }
        action: { type: string }
        resource: { type: string }
        resourceId: { type: string, format: uuid }
        reason: { type: string }
        result:
          type: string
          enum: [succeeded, failed]
        correlationId: { type: string }
        occurredAt: { type: string, format: date-time }
        beforeState: {}
        afterState: {}
    PlatformAuditEvent:
      type: object
      additionalProperties: false
      required: [id, actorType, actorId, action, resource, resourceId, reason, result, correlationId, occurredAt]
      properties:
        id: { type: string, format: uuid }
        actorType:
          type: string
          enum: [platform_administrator, tenant_user]
        actorId: { type: string, format: uuid }
        tenantId: { type: string, format: uuid, nullable: true }
        action: { type: string }
        resource: { type: string }
        resourceId: { type: string, format: uuid }
        reason: { type: string }
        result:
          type: string
          enum: [succeeded, failed]
        correlationId: { type: string }
        occurredAt: { type: string, format: date-time }
        beforeState: {}
        afterState: {}
    AccessTokenResponse:
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresIn, csrfToken]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          format: int64
          minimum: 1
        csrfToken:
          type: string
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        traceId:
          type: string